refactor: move algebra test to algebra module
